+++ Solution/Fabric.New.Operations/Traversal/Routing/TravStepParam.cs
@@ -10,8 +10,8 @@ namespace Fabric.New.Operations.Traversal.Routing {
 		public string Name { get; private  set; }
 		public Type DataType { get; private set; }
 		public bool IsGenericDataType { get; internal set; }
-		public int? Min { get; internal set; }
-		public int? Max { get; internal set; }
+		public long? Min { get; internal set; }
+		public long? Max { get; internal set; }
 		public int? LenMax { get; internal set; }
 		public string ValidRegex { get; internal set; }
 		public IList<string> AcceptedStrings { get; internal set; }
